+++ src/main/java/owl/run/modules/Transformers.java
@@ -23,23 +23,20 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static owl.ltl.rewriter.SimplifierFactory.Mode;
 import static owl.ltl.rewriter.SimplifierFactory.apply;
 
-import java.util.List;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 import owl.automaton.AutomatonUtil;
 import owl.automaton.acceptance.RabinAcceptance;
 import owl.automaton.acceptance.optimizations.AcceptanceOptimizations;
 import owl.ltl.LabelledFormula;
 import owl.run.Environment;
-import owl.run.PipelineExecutionContext;
 import owl.translations.dra2dpa.IARBuilder;
 
 public final class Transformers {
   public static final Transformer LTL_SIMPLIFIER = Transformers.fromFunction(
     LabelledFormula.class, x -> apply(x, Mode.SYNTACTIC_FIXPOINT));
-  public static final Transformer MINIMIZER =
+  public static final Transformer ACCEPTANCE_OPTIMIZATION_TRANSFORMER =
     new AcceptanceOptimizations.AcceptanceOptimizationTransformer();
-  public static final Transformer RABIN_TO_PARITY = environment -> (input, context) ->
+  public static final Transformer RABIN_TO_PARITY = environment -> (input) ->
     new IARBuilder<>(AutomatonUtil.cast(input, RabinAcceptance.class)).build();
 
   private Transformers() {
@@ -51,29 +48,12 @@ public final class Transformers {
 
   public static <K, V> Transformer.Instance instanceFromFunction(Class<K> inputClass,
     Function<K, V> function) {
-    return (object, context) -> {
+    return (object) -> {
       //noinspection ConstantConditions
       checkArgument(inputClass.isInstance(object), "Expected type %s, got type %s", inputClass,
         object == null ? null : object.getClass());
       return function.apply(inputClass.cast(object));
     };
-  }
-
-  /**
-   * Creates a {@link Transformer transformer} from a {@link OutputWriter writer} by redirecting
-   * the output to the {@link PipelineExecutionContext#metaWriter() meta writer}.
-   */
-  public static Transformer fromWriter(OutputWriter writer) {
-    return environment -> (input, context) -> {
-      writer.bind(context.metaWriter(), environment).write(input);
-      return input;
-    };
-  }
-
-  public static List<Transformer.Instance> build(List<Transformer> transformers, Environment env) {
-    return transformers.stream()
-      .map(transformer -> transformer.create(env))
-      .collect(Collectors.toUnmodifiableList());
   }
 
   public abstract static class SimpleTransformer implements Transformer.Instance, Transformer {