+++ src/main/java/owl/translations/modules/LTL2DAModule.java
@@ -19,6 +19,7 @@
 
 package owl.translations.modules;
 
+import java.io.IOException;
 import java.util.EnumSet;
 import owl.ltl.LabelledFormula;
 import owl.run.modules.ImmutableTransformerParser;
@@ -43,12 +44,12 @@ public final class LTL2DAModule {
 
   private LTL2DAModule() {}
 
-  public static void main(String... args) {
+  public static void main(String... args) throws IOException {
     PartialConfigurationParser.run(args, PartialModuleConfiguration.builder("ltl2da")
       .reader(InputReaders.LTL)
       .addTransformer(Transformers.LTL_SIMPLIFIER)
       .addTransformer(CLI)
-      .addTransformer(Transformers.MINIMIZER)
+      .addTransformer(Transformers.ACCEPTANCE_OPTIMIZATION_TRANSFORMER)
       .writer(OutputWriters.HOA)
       .build());
   }